+++ etc/blocks/enter-numbers/workflow/index.js
@@ -1,3 +1,0 @@
-module.exports = { Templates: {
-  'main': { type: 'from-file', path: require.resolve('./dist/tengo/tpl/main.plj.gz') }
-}}