+++ tools/block-tools/src/v2/model/content_conversion.ts
@@ -16,6 +16,7 @@ import {
   ContentRelativeBinary,
   ContentRelativeText
 } from '@milaboratories/pl-model-middle-layer';
+import { tryResolve } from '@milaboratories/resolve-helper';
 
 type ContentCtxFs = {
   type: 'local';
@@ -31,17 +32,6 @@ type ContentCtxUrl = {
 
 /** Describes a place relative to which any content references should be interpreted */
 export type ContentCtx = ContentCtxFs | ContentCtxUrl;
-
-function tryResolve(root: string, request: string): string | undefined {
-  try {
-    return require.resolve(request, {
-      paths: [root]
-    });
-  } catch (err: any) {
-    if (err.code !== 'MODULE_NOT_FOUND') throw err;
-  }
-  return undefined;
-}
 
 function mustResolve(root: string, request: string): string {
   const res = tryResolve(root, request);