+++ lib/node/pl-drivers/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @milaboratories/pl-drivers
 
+## 1.2.21
+
+### Patch Changes
+
+- Updated dependencies [fcbc2df]
+  - @milaboratories/ts-helpers@1.0.26
+  - @milaboratories/pl-tree@1.3.11
+  - @milaboratories/computable@2.1.6
+  - @milaboratories/pl-client@2.4.14
+
+## 1.2.20
+
+### Patch Changes
+
+- Updated dependencies [6d4cf3e]
+  - @milaboratories/pl-model-common@1.3.13
+
 ## 1.2.19
 
 ### Patch Changes